backend/requirements.txt
Django==5.2.1
django-jazzmin==3.0.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
django-cors-headers==4.7.0
pillow==11.2.1
python-decouple==3.8
psycopg2-binary==2.9.10
djangorestframework==3.15.2
gunicorn==23.0.0
django-redis
google-cloud-translate
